+++ src/main/java/com/github/mnesikos/lilcritters/configuration/ConfigLC.java
@@ -9,9 +9,8 @@ import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
 import org.zawamod.util.SpawnUtils;
 
-import static com.github.mnesikos.lilcritters.init.ModEntities.excludeBiomes;
+import static com.github.mnesikos.lilcritters.init.ModEntities.excludeBiomesByType;
 import static com.github.mnesikos.lilcritters.init.ModEntities.getAllBiomesOf;
-import static org.zawamod.util.SpawnUtils.mergeBiomes;
 
 @Config(modid = Ref.MODID, name = "LilCritters-" + Ref.VERSION +"/"+ Ref.MODID)
 @Config.LangKey("config.lilcritters.title")
@@ -31,26 +30,9 @@ public class ConfigLC {
     @Config.RangeDouble(min = 0.0, max = 1.0)
     public static double treeSeedDropsChance = 0.08;
 
-    @Config.Comment("Spawns Configuration")
+    @Config.Comment({"Spawns Configuration", "LC'S SPAWN SETTINGS ARE CURRENTLY DISABLED HERE. Please use ZAWA's config file to edit LC's spawn settings."})
     public static final Spawns spawns = new Spawns();
-    public static class Spawns {
-        @Config.Name("Tree Squirrels")
-        public final SpawnSettings tree_squirrel = new SpawnSettings(30, 1, 2, SpawnUtils.getBiomeIDS(getAllBiomesOf(BiomeDictionary.Type.FOREST)));
-        @Config.Name("Box Turtles")
-        public final SpawnSettings box_turtle = new SpawnSettings(20, 1, 1, SpawnUtils.getBiomeIDS(excludeBiomes(getAllBiomesOf(BiomeDictionary.Type.FOREST, BiomeDictionary.Type.SANDY, BiomeDictionary.Type.PLAINS), BiomeDictionary.Type.COLD)));
-        @Config.Name("Tufted Deer")
-        public final SpawnSettings tufted_deer = new SpawnSettings(10, 1, 1, SpawnUtils.getBiomeIDS(getAllBiomesOf(BiomeDictionary.Type.HILLS, BiomeDictionary.Type.FOREST)));
-        @Config.Name("Skunks")
-        public final SpawnSettings skunk = new SpawnSettings(10, 1, 2, SpawnUtils.getBiomeIDS(getAllBiomesOf(BiomeDictionary.Type.FOREST, BiomeDictionary.Type.PLAINS, BiomeDictionary.Type.SAVANNA)));
-        @Config.Name("Opossum")
-        public final SpawnSettings opossum = new SpawnSettings(20, 1, 1, SpawnUtils.getBiomeIDS(excludeBiomes(getAllBiomesOf(BiomeDictionary.Type.SAVANNA, BiomeDictionary.Type.JUNGLE, BiomeDictionary.Type.FOREST, BiomeDictionary.Type.PLAINS, BiomeDictionary.Type.MOUNTAIN, BiomeDictionary.Type.SWAMP), BiomeDictionary.Type.COLD)));
-        @Config.Name("Dwarf Crocodile")
-        public final SpawnSettings dwarf_crocodile = new SpawnSettings(10, 1, 1, SpawnUtils.getBiomeIDS(excludeBiomes(getAllBiomesOf(BiomeDictionary.Type.RIVER, BiomeDictionary.Type.SWAMP, BiomeDictionary.Type.WET, BiomeDictionary.Type.SAVANNA, BiomeDictionary.Type.FOREST), BiomeDictionary.Type.COLD)));
-        @Config.Name("Small Clawed Otter")
-        public final SpawnSettings small_clawed_otter = new SpawnSettings(20, 4, 6, SpawnUtils.getBiomeIDS(getAllBiomesOf(BiomeDictionary.Type.RIVER, BiomeDictionary.Type.BEACH)));
-        @Config.Name("Banded Penguin")
-        public final SpawnSettings banded_penguin = new SpawnSettings(15, 4, 8, SpawnUtils.getBiomeIDS(getAllBiomesOf(BiomeDictionary.Type.BEACH)));
-    }
+    public static class Spawns {}
 
     public static class SpawnSettings {
         @Config.Name("Spawn Chance")